fix error in deepseek when using allgather
